Move utilities into form config
